VGG16/requirements.txt
tensorflow==2.15.0
numpy==1.19.5
requests==2.26.0
Pillow==8.2.0
